feat(312): C2 extract RouterCapExceeded → runtime/errors.py (god-file decomposition)
FP-0044 C-series stage 2. Extracts the `RouterCapExceeded` exception verbatim
from `reyn.runtime.session` (the god-file) into its own `reyn.runtime.errors`
module so the runtime exception type lives apart from the Session god-class.

- Pure move, byte-identical class body (no behavior change, no class rename).
- `runtime/errors.py` docstring is current-state only (what the module IS);
  this refactor narrative lives here in the commit, per the C-series docstring
  discipline established in C1 (#1790 R1).
- Clean break, no re-export shim: all 12 consumer files repointed
  `from reyn.runtime.session import RouterCapExceeded`
  → `from reyn.runtime.errors import RouterCapExceeded` (mixed-import lines
  split so `Session` / `_new_chain_id` stay sourced from session).
- session.py imports the type back from errors.py for its own internal raises.
- Repo-wide 4-ref-class straggler grep (dotted-import / dotted-literal /
  segment-path / docstring+comment) = 0 references to the old location.
- errors.py imports nothing from reyn (zero cycle risk); one-directional
  session → errors.
