+++ tests/test_budget_gateway_invariants.py
@@ -17,8 +17,8 @@ from reyn.core.events.events import EventLog
 from reyn.llm.model_resolver import ModelResolver
 from reyn.llm.pricing import TokenUsage
 from reyn.runtime.budget.budget import BudgetTracker, CostConfig
+from reyn.runtime.errors import RouterCapExceeded
 from reyn.runtime.services.budget_gateway import BudgetGateway
-from reyn.runtime.session import RouterCapExceeded
 
 # ---------------------------------------------------------------------------
 # Helpers
