+++ tests/test_safety_phase2_wiring.py
@@ -25,8 +25,9 @@ import pytest
 from reyn.config import LoopConfig, OnLimitConfig, SafetyConfig
 from reyn.core.kernel.runtime import LoopLimitExceededError, OSRuntime
 from reyn.runtime.budget.budget import BudgetTracker, CostConfig
+from reyn.runtime.errors import RouterCapExceeded
 from reyn.runtime.limits.limit_handler import reset_run_extensions
-from reyn.runtime.session import RouterCapExceeded, Session
+from reyn.runtime.session import Session
 from reyn.schemas.models import Phase, Skill, SkillGraph
